Rename core.data -> core.proxies
