+++ registrar/apps/api/v1/mixins.py
@@ -21,9 +21,9 @@ from rest_framework.status import (
 )
 
 from registrar.apps.core import permissions as perms
-from registrar.apps.core.data import DiscoveryProgram
 from registrar.apps.core.filestore import get_enrollment_uploads_filestore
 from registrar.apps.core.jobs import start_job
+from registrar.apps.core.proxies import DiscoveryProgram
 from registrar.apps.enrollments.data import (
     write_course_run_enrollments,
     write_program_enrollments,
